whisper_backend/pyproject.toml
[tool.poetry]
name = "whisper-backend"
version = "0.1.0"
description = ""
authors = ["Veter-Alex <Vet-an@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
psycopg2-binary = "^2.9.10"
loguru = "^0.7.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py310']